Changed database: added rentals table and corrected other table

+++ database/migrations/2025_09_16_172110_create_car_models_table.php
@@ -15,7 +15,7 @@ return new class extends Migration {
             $table->string('title');
             $table->enum('engine_fuel_type', ['gasoline', 'diesel', 'hybrid', 'electric']);
             $table->enum('transmission', ['AT', 'MT']);
-            $table->integer('cost_per_day');
+            $table->float('cost_per_day');
             $table->integer('seat_quantity');
             // Внешние ключи
             $table
